pyproject.toml
[tool.poetry]
name = "mono"
version = "0.1.0"
description = ""
authors = ["Dasieneous <dasieneous@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
aiosqlite = "^0.19.0"
asyncpg = "^0.27.0"
alembic = "^1.10.4"
sqlalchemy = "^2.0.12"
greenlet = "^2.0.2"
psycopg2 = "^2.9.6"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
factory-boy = "^3.2.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
pytest-factoryboy = "^2.5.1"
uvicorn = "^0.22.0"
fastapi-utils = { git = "https://github.com/dasieneous/fastapi-utils.git"}



[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
